3D-Human-Pose-Estimation-main"
This studies delves into the development and evaluation of forms designed to reveal and 
estimate human postures using deep mastering strategies. Our primary cognizance is on the 
GraFormer model, a Graph Convolution Transformer specifically designed for 3D Pose 
Estimation. It's difficult to as it should be predict human postures from pictures and movement 
captures. TensorFlow and Keras were used inside the production, training, and contrast of model 
architectures primarily based on pose facts. Excellent posture prediction accuracy is tested 
through the consequences, which are further corroborated with the aid of using wealthy 
visualizations. The mission's dreams, methodology, conclusions, and findings are all 
significantly protected on this record.